Rename CoordinateExport... to ConfigurationExport...
